requirements.txt
asgiref==3.4.1
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==36.0.1
defusedxml==0.7.1
diff-match-patch==20200713
django-allauth==0.47.0
django-crispy-forms==1.14.0
django-import-export==2.7.1
django-sendgrid-v5==1.2.0
Django==4.0.1
docopt==0.6.2
et-xmlfile==1.1.0
gunicorn==20.1.0 
google-api-core==2.4.0
google-api-python-client==2.36.0
google-auth-httplib2==0.1.0
google-auth==2.3.3
googleapis-common-protos==1.54.0
httplib2==0.20.2
idna==3.3
import-export==0.2.67.dev6
MarkupPy==1.14
oauthlib==3.1.1
odfpy==1.4.1
openpyxl==3.0.9
Pillow==9.0.0
pip==21.1.2
pipreqs==0.4.11
protobuf==3.19.3
psycopg2==2.9.3
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
python-http-client==3.3.5
python3-openid==3.2.0
PyYAML==6.0
requests-oauthlib==1.3.0
requests==2.27.1
rsa==4.8
sendgrid==6.9.5
setuptools==57.0.0
six==1.16.0
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
tablib==3.1.0
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
wheel==0.36.2
xlrd==2.0.1
xlwt==1.3.0
yarg==0.1.9

